Frontend/requirements.txt
streamlit
requests
pandas
plotly
Pillow
yfinance
